small fix
[fix]: makefile no longer relink
[fix]: change strlen to use size_t

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: mjuncker <mjuncker@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/24 17:58:17 by mjuncker          #+#    #+#             */
-/*   Updated: 2024/11/06 16:40:15 by mjuncker         ###   ########.fr       */
+/*   Updated: 2024/11/11 13:41:00 by mjuncker         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (s[len])
